+++ modules/engine/src/cpp/AddInternalGateways.cpp
@@ -26,6 +26,7 @@
 #include "core_Gateway.hpp"
 #include "error_manager_Gateway.hpp"
 #include "memory_manager_Gateway.hpp"
+#include "data_structures_Gateway.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
@@ -49,6 +50,7 @@ addInternalGateways(Evaluator* eval)
     CoreGateway(eval, L"");
     ErrorManagerGateway(eval, L"");
     MemoryManagerGateway(eval, L"");
+    DataStructuresGateway(eval, L"");
 }
 //=============================================================================
 }